+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/application/usecase/ProcessFileUseCase.java
@@ -1,12 +1,12 @@
 package de.muenchen.oss.swim.dms.application.usecase;
 
 import de.muenchen.oss.swim.dms.application.port.out.DmsOutPort;
+import de.muenchen.oss.swim.dms.application.usecase.helper.TargetResolverHelper;
 import de.muenchen.oss.swim.dms.configuration.DmsMeter;
 import de.muenchen.oss.swim.dms.configuration.SwimDmsProperties;
 import de.muenchen.oss.swim.dms.domain.exception.DmsException;
 import de.muenchen.oss.swim.dms.domain.helper.DmsMetadataHelper;
 import de.muenchen.oss.swim.dms.domain.helper.PatternHelper;
-import de.muenchen.oss.swim.dms.domain.model.DmsResourceType;
 import de.muenchen.oss.swim.dms.domain.model.DmsTarget;
 import de.muenchen.oss.swim.dms.domain.model.UseCase;
 import de.muenchen.oss.swim.dms.domain.model.UseCaseType;
@@ -21,11 +21,9 @@ import de.muenchen.oss.swim.libs.handlercore.domain.model.FileEvent;
 import de.muenchen.oss.swim.libs.handlercore.domain.model.Metadata;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -42,6 +40,7 @@ public class ProcessFileUseCase implements ProcessFileInPort {
     private final FileEventOutPort fileEventOutPort;
     private final DmsMetadataHelper dmsMetadataHelper;
     private final PatternHelper patternHelper;
+    private final TargetResolverHelper targetResolverHelper;
     private final DmsMeter dmsMeter;
 
     @Override
@@ -67,7 +66,7 @@ public class ProcessFileUseCase implements ProcessFileInPort {
                 targetResource = useCase.getType();
             }
             // get target coo
-            final DmsTarget dmsTarget = this.resolveTargetCoo(targetResource, metadata, useCase, file);
+            final DmsTarget dmsTarget = this.targetResolverHelper.resolveTargetCoo(targetResource, metadata, useCase, file);
             log.debug("Resolved dms target: {}", dmsTarget);
             // get ContentObject name and subject
             final String contentObjectName = this.patternHelper.applyPattern(useCase.getContentObject().getFilenameOverwritePattern(), file.getFileName(),
@@ -168,94 +167,6 @@ public class ProcessFileUseCase implements ProcessFileInPort {
     }
 
     /**
-     * Resolve target coo for useCase.
-     * {@link UseCaseType}
-     *
-     * @param resourceType Target type the coo is resolved for.
-     * @param metadata Parsed metadata file.
-     * @param useCase The use case.
-     * @param file The file to resolve the coo for.
-     * @return The resolved coo.
-     */
-    protected DmsTarget resolveTargetCoo(final UseCaseType resourceType, final Metadata metadata, final UseCase useCase, final File file)
-            throws MetadataException {
-        return switch (useCase.getCooSource().getType()) {
-        case METADATA_FILE -> this.resolveMetadataTargetCoo(resourceType, metadata, useCase);
-        case FILENAME -> {
-            if (Strings.isBlank(useCase.getCooSource().getFilenameCooPattern())) {
-                throw new IllegalArgumentException("Filename coo pattern is required");
-            }
-            final String targetCoo = this.patternHelper.applyPattern(useCase.getCooSource().getFilenameCooPattern(), file.getFileName(), metadata);
-            yield new DmsTarget(targetCoo, useCase.getContext());
-        }
-        case FILENAME_MAP -> {
-            // find first matching target coo from map
-            final String targetCoo = useCase.getCooSource().getFilenameToCoo().entrySet().stream()
-                    .filter(i -> Pattern.compile(i.getKey(), Pattern.CASE_INSENSITIVE).matcher(file.getFileName()).find())
-                    .findFirst()
-                    .map(Map.Entry::getValue)
-                    .orElseThrow(() -> new IllegalStateException("No matching filename map entry configured."));
-            yield new DmsTarget(targetCoo, useCase.getContext());
-        }
-        case FILENAME_NAME -> this.resolveTargetCooViaName(useCase.getType().getTarget(), metadata, useCase, file);
-        case STATIC -> new DmsTarget(useCase.getCooSource().getTargetCoo(), useCase.getContext());
-        case OU_WORK_QUEUE -> new DmsTarget(null, useCase.getContext());
-        };
-    }
-
-    /**
-     * Resolve DmsTarget via metadata file.
-     *
-     * @param resourceType The type of the use case used for resolving.
-     * @param metadata Parsed metadata file.
-     * @param useCase UseCase of the file.
-     * @return Resolved DmsTarget.
-     */
-    protected DmsTarget resolveMetadataTargetCoo(final UseCaseType resourceType, final Metadata metadata, final UseCase useCase) throws MetadataException {
-        // validate metadata provided
-        if (metadata == null) {
-            throw new MetadataException("Target coo via metadata file: Metadata is required");
-        }
-        // extract coo and username from metadata
-        final DmsTarget metadataTarget = switch (resourceType) {
-        case INBOX_CONTENT_OBJECT -> dmsMetadataHelper.resolveInboxDmsTarget(metadata);
-        case PROCEDURE_INCOMING -> dmsMetadataHelper.resolveIncomingDmsTarget(metadata);
-        case METADATA_FILE -> throw new IllegalStateException("Target type metadata needs to be resolved to other types");
-        };
-        // combine resolves target with use case
-        return this.combineDmsTargetWithUseCase(metadataTarget, useCase);
-    }
-
-    /**
-     * Resolve target coo via dms object name.
-     *
-     * @param resourceType The resource type of the target.
-     * @param metadata The metadata used for resolving the name pattern.
-     * @param useCase The use case to resolve the target for.
-     * @param file The file to resolve the target for.
-     * @return Dms target resolved via dms object name.
-     */
-    protected DmsTarget resolveTargetCooViaName(final DmsResourceType resourceType, final Metadata metadata, final UseCase useCase, final File file) {
-        // validate required use case properties
-        if (Strings.isBlank(useCase.getCooSource().getFilenameNamePattern())) {
-            throw new IllegalArgumentException("DMS target coo via object name: Filename name pattern is required");
-        }
-        if (Strings.isBlank(useCase.getContext().getUsername())) {
-            throw new IllegalStateException("DMS target coo via object name: Username is required");
-        }
-        // resolve lookup name
-        final String objectName = this.patternHelper.applyPattern(useCase.getCooSource().getFilenameNamePattern(), file.getFileName(), metadata);
-        // search for object name
-        final DmsTarget requestContext = new DmsTarget(null, useCase.getContext());
-        final List<String> coos = this.dmsOutPort.findObjectsByName(resourceType, objectName, requestContext);
-        if (coos.size() != 1) {
-            throw new IllegalStateException(
-                    String.format("DMS target coo via object name: Found %d instead of exactly one object for pattern %s", coos.size(), objectName));
-        }
-        return new DmsTarget(coos.getFirst(), useCase.getContext());
-    }
-
-    /**
      * Resolve dms target resource type from metadata file.
      *
      * @param metadata Parsed metadata file.
@@ -308,26 +219,5 @@ public class ProcessFileUseCase implements ProcessFileInPort {
                         i.getKey().replaceFirst("^" + swimDmsProperties.getMetadataSubjectPrefix(), ""),
                         i.getValue()))
                 .collect(Collectors.joining("\n"));
-    }
-
-    /**
-     * Combine resolved DmsTarget with UseCase values.
-     * Checks that only one of both defines job oe and position values.
-     *
-     * @param dmsTarget Resolved target.
-     * @param useCase UseCase the target was resolved for.
-     * @return Combined DmsTarget.
-     * @throws IllegalStateException If both inputs define job oe or position.
-     */
-    protected DmsTarget combineDmsTargetWithUseCase(final DmsTarget dmsTarget, final UseCase useCase) {
-        final boolean dmsTargetHasJob = Strings.isNotBlank(dmsTarget.getJoboe()) || Strings.isNotBlank(dmsTarget.getJobposition());
-        final boolean useCaseHasJob = Strings.isNotBlank(useCase.getContext().getJoboe()) || Strings.isNotBlank(useCase.getContext().getJobposition());
-        if (dmsTargetHasJob && useCaseHasJob) {
-            throw new IllegalStateException("Resolve dms target: Job oe and position defined via resolve and via use case not allowed");
-        }
-        if (dmsTargetHasJob) {
-            return dmsTarget;
-        }
-        return new DmsTarget(dmsTarget.getCoo(), dmsTarget.getUsername(), useCase.getContext().getJoboe(), useCase.getContext().getJobposition());
     }
 }
